# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LaraSol API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/farmers": {
            "get": {
                "tags": [
                    "Farmers"
                ],
                "description": "Get all Farmers",
                "operationId": "getFarmersList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Store new Farmer",
                "description": "Returns farmer data",
                "operationId": "storeFarmers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFarmerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farmer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/farmers/{id}": {
            "get": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Get farmer information",
                "description": "Returns farmer data",
                "operationId": "getFarmerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Farmer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farmer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Update existing farmer",
                "description": "Returns updated farmer data",
                "operationId": "updateFarmers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "County id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFarmerRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farmer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Farmers"
                ],
                "summary": "Delete existing farmer",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteFarmer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Farmer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreFarmerRequest": {
                "title": "Store Farmer Form request",
                "required": [
                    "county_name"
                ],
                "properties": {
                    "national_id": {
                        "description": "National Id no.",
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Doe"
                    },
                    "gender": {
                        "description": "Gender male or female",
                        "type": "string",
                        "readOnly": "true",
                        "example": "male"
                    },
                    "dob": {
                        "description": "Date of Birth",
                        "type": "string",
                        "format": "date-time",
                        "example": "2019-02-25"
                    },
                    "email": {
                        "description": "Farmer unique email address",
                        "type": "string",
                        "format": "email",
                        "example": "farmer@gmail.com"
                    },
                    "phone": {
                        "description": "Farmer phone number",
                        "type": "string",
                        "example": "254716365943"
                    },
                    "status": {
                        "description": "Registration status",
                        "type": "string",
                        "example": "Registered"
                    },
                    "county": {
                        "description": "Farm County Name",
                        "type": "string",
                        "example": "Nyeri"
                    },
                    "ward": {
                        "description": "Farm ward name",
                        "type": "string",
                        "example": "MATHIOYA"
                    },
                    "village": {
                        "description": "Farm village name",
                        "type": "string",
                        "example": "Kariguini"
                    },
                    "farm_type": {
                        "description": "Type of Farming",
                        "type": "string",
                        "example": "Subsistence"
                    },
                    "produce": {
                        "description": "List of farm produce",
                        "type": "string",
                        "example": "['Mangoes','Cashew Nut']"
                    },
                    "farm_house": {
                        "description": "Image of farm house",
                        "type": "string",
                        "example": "img_20230103_220125.jpg"
                    },
                    "terms": {
                        "description": "Terms and Condition",
                        "type": "string",
                        "example": "Accepted"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "StoreFarmerRequest"
                }
            },
...
            "UpdateFarmerRequest": {
                "title": "Store Farmer Form request",
                "required": [
                    "county_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "national_id": {
                        "description": "National Id no.",
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Doe"
                    },
                    "gender": {
                        "description": "Gender male or female",
                        "type": "string",
                        "readOnly": "true",
                        "example": "male"
                    },
                    "dob": {
                        "description": "Date of Birth",
                        "type": "string",
                        "format": "date-time",
                        "example": "2019-02-25"
                    },
                    "email": {
                        "description": "Farmer unique email address",
                        "type": "string",
                        "format": "email",
                        "example": "farmer@gmail.com"
                    },
                    "phone": {
                        "description": "Farmer phone number",
                        "type": "string",
                        "example": "254716365943"
                    },
                    "status": {
                        "description": "Registration status",
                        "type": "string",
                        "example": "Registered"
                    },
                    "county": {
                        "description": "Farm County Name",
                        "type": "string",
                        "example": "Nyeri"
                    },
                    "ward": {
                        "description": "Farm ward name",
                        "type": "string",
                        "example": "MATHIOYA"
                    },
                    "village": {
                        "description": "Farm village name",
                        "type": "string",
                        "example": "Kariguini"
                    },
                    "farm_type": {
                        "description": "Type of Farming",
                        "type": "string",
                        "example": "Subsistence"
                    },
                    "produce": {
                        "description": "List of farm produce",
                        "type": "string",
                        "example": "['Mangoes','Cashew Nut']"
                    },
                    "farm_house": {
                        "description": "Image of farm house",
                        "type": "string",
                        "example": "img_20230103_220125.jpg"
                    },
                    "terms": {
                        "description": "Terms and Condition",
                        "type": "string",
                        "example": "Accepted"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UpdateFarmerRequest"
                }
            },
...
            "Farmer": {
                "required": [
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "national_id": {
                        "description": "National Id no.",
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "John"
                    },
                    "last_name": {
                        "type": "string",
                        "maxLength": 32,
                        "example": "Doe"
                    },
                    "gender": {
                        "description": "Gender male or female",
                        "type": "string",
                        "readOnly": "true",
                        "example": "male"
                    },
                    "dob": {
                        "description": "Date of Birth",
                        "type": "string",
                        "format": "date-time",
                        "example": "2019-02-25"
                    },
                    "email": {
                        "description": "Farmer unique email address",
                        "type": "string",
                        "format": "email",
                        "example": "farmer@gmail.com"
                    },
                    "phone": {
                        "description": "Farmer phone number",
                        "type": "string",
                        "example": "254716365943"
                    },
                    "status": {
                        "description": "Registration status",
                        "type": "string",
                        "example": "Registered"
                    },
                    "county": {
                        "description": "Farm County Name",
                        "type": "string",
                        "example": "Nyeri"
                    },
                    "ward": {
                        "description": "Farm ward name",
                        "type": "string",
                        "example": "MATHIOYA"
                    },
                    "village": {
                        "description": "Farm village name",
                        "type": "string",
                        "example": "Kariguini"
                    },
                    "farm_type": {
                        "description": "Type of Farming",
                        "type": "string",
                        "example": "Subsistence"
                    },
                    "produce": {
                        "description": "List of farm produce",
                        "type": "string",
                        "example": "['Mangoes','Cashew Nut']"
                    },
                    "farm_house": {
                        "description": "Image of farm house",
                        "type": "string",
                        "example": "img_20230103_220125.jpg"
                    },
                    "terms": {
                        "description": "Terms and Condition",
                        "type": "string",
                        "example": "Accepted"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
        }
    }
}